+++ main/comp_net/src/asteroid_simulation/simulation_client.cpp
@@ -153,7 +153,7 @@ void SimulationClient::ReceivePacket(const asteroid::Packet* packet)
         case asteroid::PacketType::START_GAME:
         {
             const auto* startGamePacket = static_cast<const asteroid::StartGamePacket*>(packet);
-            unsigned long long startingTime = ConvertFromBinary<unsigned long long>(startGamePacket->startTime);
+            unsigned long startingTime = ConvertFromBinary<unsigned long>(startGamePacket->startTime);
             gameManager_.StartGame(startingTime);
             break;
         }
